app/requirements.txt
# Servidor API
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
aiofiles==23.2.1
python-multipart==0.0.9

# Cliente vector DB
qdrant-client==1.8.2

# Embeddings faciales (GPU)
insightface==0.7.3
onnxruntime-gpu==1.16.3

# Utilidades
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
tqdm==4.66.5
rich==13.8.1
orjson==3.10.7
